Fix migration screw up.
This appears to be the only point where I've made a huge mistake: I
thought what we were storing here was an unknown obj in the instance
record. Turns out that actually what the original code wanted to do was
point out that it had unknown validity, which we've moved in the new
irep to be two bools rather than an object.

Correct this by, upon free, marking an object as having unknown validity
(could be a nondet bool).