Twiddle with a regression test.
So, currently this test is failing by saying the solver failed, after
constructing the formula. Immediately observe that if you bump up the
exploration depth with an older version of ESBMC (to say, 6000), the
same thing happens.

So whatever is happening is due to a change in the amount of instruction
exploration going on. From a brief look, that may be because of
simplification of array_of's, which is an additional thing I've
implemented. So, fewer instructions are likely being executed in the
current arrangement.

Because of this, just reduce the exploration depth. The bug isn't
introduced by me, so just avoid it for now, it's not something I'm
persuing atm.