Handle derefencing of invalid symbols.
i.e., follow previous behaviour, of not encoding a check if we're
dereferencing an invalid object. Check for this not by looking at some
arbitary field of an irep, but by looking at the symbol prefix to see if
it was generated by dereferencet.

This is about as horrible as the arbitary-field situation from before.
The correct fix is to keep a set of symbol names in a dereferencet, and
to make the dereferencet persistent.