Fix an intrinsics type assertion.
A call to get_thread_start_data is getting some kind of additional type
data in a pointer being flattened to empty. Don't know why, but it's
still valid, just the types don't compare equal.

So, use base_type_eq to assert that the types are at least the same
shape, if not exactly the same type.