Correct overflow simplification.
Just realised that for the operand-is-constant case, the operand will
already have been truncated into a smaller integer. So there's no way to
test for overflow; the available operand will always fit in the
datatype.

So now detect overflow by simplifying the operand twice, with different
type widths, and see whether or not the results differ.