Move insn code to irep 2. (broken build).
So, this moves the code and guard field in an instructiont to irep 2.
Hurrah, the witch is dead.

Everything in goto-symex and pointer-analysis compiles. However,
goto-programs is going to be a trainwreck, because of all the
moving-from-exprs-to-insns stuff. For example,
goto_convertt::finish_gotos appears to perform partial conversions :|

Still, note the change to goto_program_template.h. It looks like the
whole goto-programs-are-code-generic attempt with templating up programs
is rubbish.

(This also whips some dead code from goto_convert.cpp, which was the
point where I realised this is going to have to be over several
commits.)