Attempt to migrate string instru., mostly fail
A bunch of things I haven't encountered before, but most importantly a
"buffer_size" irep that we've never supported at all. This can't have
worked in the past, and it won't work now.

Just disable portions of code that I don't understand; they can be
re-enabled in the future when there's a buffer size irep.