Absolutely massacre string abstraction.
This is going to turn out to be a trainwreck at runtime. There's a bunch
of things that are dodgy in this, including one point (see comments)
where it turns out a typet was being implicitly casted to being an
exprt. Which is unpleasent to say the least.

On the surface it seems I've been able to maintain a 1 <=> 1 mapping of
old irep to new irep conversion (sans the "buffer_size" irep), but I'd
bet good money I've accidentally screwed something up in here.