Goto code irep and migration.
This is code to encapsulate what I think is a user provided goto. It
just stores a target name; it could probably be stripped out of irep
before being turned into a goto instruction, but that's not how the
conversion code has chosen to do it.

(Looks like that's because it connects the goto to the target
 instruciton in some kind of final linking stage, which makes a lot of
 sense)