Kill buffer_size irep conversion code.
I discovered a while ago that the reason why this never turned up at the
solver in the past was that string abstraction rewrote buffer_size ireps
into something else (which I'd disabled at the time). Kill this
conversion routine though because a) it's broken, and b) it causes
warnings, which now causes errors.