Fix some type mismatching when using byte extract.
Byte extract _always_ produces one byte output. However in this
situation, we just expect it to produce a value of whatever size
to_type is. Which is bad.

With irep changes, we end up with Z3 throwing an error because we're
making an equality between a 32 bit int and the 8 bit result of
byte_extract.

Work around this by upcasting the byte from byte_extract to whatever
type we expect. This is 100% broken; but without updating the pointer
gunge implementation (a big behaviour change) there's no way of fixing
that.

see #126 which records this.