Allow reachability_treet to be re-used.
So, a change in behaviour I introduced caused reachability_treets that
had run to completion to not automatically restart if they were invoked
again. This isn't a bad thing, because it ensures we don't silently
restart exactly the same exploration unexpectedly.

However k-induction repeatedly uses the same bmct object to perform
different steps; which really need to be able to restart exploration.
So, have a new bmc run set up a new reachability_treet run.

Also insert an assertion that a reachability tree obj has been set up
before we start executing it.