Work around a Z3 typing issue.
So for 32_floppy and 33_blah, we're currently experiencing a situation
where two structs are either being compared or used in an if-then-else
case, which /can/ be successful because they're either compatible
structs or have a compatible prefix.

However, Z3 objects to this, apparently on the basis that they have to
be _the_ same sort, both in name and structure.

So, if we approve a dereference in dereference_type_compare because the
types are sufficiently similar, insert a typecast to make sure that the
object type is casted to the dereference type.