Deal with fallout of inserting more deref casts
Now the objects returned from dereference_type_compare may very well be
wrapped in some typecast; if that's the case, unpack index2t's from it
before updating the offset and inserting a bounds check.

Otherwise we fall through and insert an assertion that the pointer
offset is never nonzero, which breaks a /lot/ of things.