Don't miss initializing an operand to a union.
We feed an array of union members to Z3 in a tuple; however by iterating
over all the member operands, we're a victim to the "the number of union
operands is entirely nondeterminstic" problem. Some members of the array
of members don't get initialized as a result, leading to boom.

Fix this by iterating over the types expected, that way all fields get
either a union operand or a fresh constant.