Work around a constant array_of situation.
If it's infinitely sized, we no longer support pumping a constant
array_of into SMT, because SMT doesn't support it anyway (although Z3
itself does). See: revision history of z3_conv.cpp.

For the pointer tracking data arrays this is fine because we can hack
some C++ to make assumptions that initialize a free array. However,
we've also got some infinitely sized arrays in pthread at the moment,
where this breaks horribly.

So, have on thread creation, the field of the tracking array written to
ensure that it's the right value.

But that's not enough - the simplifier simplifies the write of false to
an array_of false's  to an array_of false's. Which is right for all
other situations except where we have an infinite sized array (and SMT
not supporting it). So, add a hack to not simplify that away.