Use boost multi-index container for bv cache.
So I want two things: to be able to use the bv cache like a normal hash
map that we've been using so far, and to be able to quickly erase a
bunch of entries from it when we pop a z3 context.

So have a cache entry contain a level, representing what push/pop level
the entry is on; this is then indexed in order, so that we can wipe out
a whole level in the future.