Seriously juggle the way bmct works.
For runtime-encoded SMT, we need to have a z3_convt that has the
lifetime of the model check (and isn't part of RT, as that shouldn't
care what target it's getting).

Handle this by having a z3_convt in bmct, and if the appropriate option
is enabled, hand RT a target to store to that contains a reference to
that object.

Quite a few of the changes here are structural, in that the symex object
becomes a pointer, as it's too much pain to initialize it as a member of
the bmct class.