Add two options for features that use symex-smt.
These are things that check the guards at particular points to see
whether or not we should be continuing exploration. Both the thread
state guard, and the symbolic execution guard.

Also add a check making us abort if the user specifies one of these
options but doesn't specif --smt-during-symex. We could enable it
implicitly, but I prefer the user acknowledging what they're doing for
this.