Define a routine for asking solver questions.
The use case is where we have a guard, and we want to see if it's true,
false or unknown. We push and pop a context around working with it to
ensure we don't mess up the existing formula encoded.

First, test the formula (q | !q), which should always be satisfiable
(and true). We can then inspect which of not q or !q is true, after
which we have an initial result. Then we assume that that literal is
false, and see whether the other one becomes true when we check
satisfiabilty again. If it does, the question can have any answer, so
the result is unknown; if not we can extract whether it's true or false.