"long long" ints aren't 128 bits on 64 bit archs.
Well, perhaps it is on some, but it's messing with the slightly
more restrictive refactored constant storage goo, which shuttles things
through 64 bit ints, and thus asserts that integers are 64 bits or less.

GCC on iridis reports that a long long integer is 64 bits. long long
should thus be 64 bit there in ESBMC.

This isn't going on master because a) it's due to the irep refactor, and
b) kinduction.