Don't call failed symbols.
We did this in the past; however, it turns out there's a second,
additional form of symbol name that indicates we've run into a failed
symbol.

I suspect in the future I'll arrange a situation where there are
multiple symbol2t subclasses that contain additional information; like
renaming and failed symbol status.