Start work on pulling down a static libc for esbmc
So; currently the ESBMC build situation is a bit sucky. This is because
we have a few options:

  1) Dynamic
  2) Compat
  3) Static
  4) Windows

With the following problems:

  1) libstdc++ dependancies, Z3 shared object requirement, latest glibc
     tacks a symbol version on memcpy because Drepper;
  2) Slower; fewer dependancy issues, but Z3 so still needs to be
     present. glibc is still a pain.
  3) None of the above problems; however the only libc's available to me
     aren't compiled to work with Linux < 2.6.32
  4) Windows

One of the solutions to this is to take option 3 and link in a libc that
isn't so limited. But then there's the problem that it's unreasonable
for normal developers to compile it themselves, and I'm not putting it
in the esbmc repository (it's a binary!).

So instead have it pulled down from my website and verified with a pgp
signature. This also allows for an updated version to be fetched if the
user deletes the version of libc.a they've downloaded. Note that this is
only ever invoked if they build "iridis-esbmc".

This also imports my pgp key into the scripts dir in a gpg keyring file.
Wooo, pgp!