Destroy the universe with even more templates.
So this is my answer to supporting hierarchy and the ability to add on
members. We need to be able to add members arbitarily, and re-generate
the templated virtual methods arbitarily (so that an intermediate types
convert_smt method is usable).

The obstacle to this is that we have to pass down constructor arguments
to something we subclass. If we want to generate new members or
reinstanciate the virtual methods use the "type" template, there's no
good way to pass down arguments to what we're subclassing.

So instead, seperate the two lumps of templates into a template that
generates the virtual methods, and one that adds a new member. Because
we're passing the most derived class down to the template, we can take a
reference to it and know the object layout in the templates anyway.

Because this leads to inhereting from type2t many times (for the virtual
methods to end up in the right place), virtual inheretance has to occur,
which also causes my face to fall off. Unpleasent, but necessary.

This copies all the esbmct::type body implementations to type_methods;
the only difference in the copy+paste is the introduction of two dynamic
casts instead of static casts, due to G++ apparently not having enough
information to statically downcast a reference to the derived type.
(Apparently RTTI gives it that information).

Next up, moving all type classes to this form, and introducing
hierarchy. And then being appalled at all the extra template complexity.