struct/union type sharing; revise templating view
This patch fixes up some functions to be part of a greater
struct_union_type2t, and makes struct_type2t and union_type2t subclass
it. Hurrah.

Notably is that I've discovered the previous idea of having both the
subclass and a new template of virtual methods merged together doesn't
work -- C++ wants a final override of the virtual methods between each
class (struct_union_type2t, and the new type_methods template) to be
created _in_ the new class (i.e. struct_type2t, union_type2t).
Otherwise, it can't work out which class to take each virtual method
from.

So fix this by only ever instanciating type methods in the final most
derived type. It turns out that before I thought this wouldn't work
because it'd break the convert_smt despatching, but as it happens that
isn't the case, it's a different piece of polyporphism handling that.

So yeah; success. Worrying that a number of things are having to turn
into dynamic casts though.