Revert attempt to move to split templates.
This reverts commit 5a056864f1d2800a5015a5ec9b312de5ad243511.
This reverts commit 659a7ba339c16e5b85d3d3ccf1485f974ca07341.

So it turns out that this mechanism just doesn't work. Perhaps it does
in theory, but i'm ending up with the vtables for expr_data and
expr_methods being unmergeably different, leading to constant_int2t
having one vtable for itself, and then an expr2t subobject that has its
own vtable. When stored in an expr2tc, the cast to expr2t* takes a ptr
to the internal object. When it gets freed, boom.

So, multiple inheretance has killed this, and I don't know why. This
didn't happen to type2t because as it turns out they never get freed
(and are permenantly cached).

So there's a better situation I've prototyped: pass down to expr_methods
or whatever a type and a /member offset pointer/. That way, we can
construct a _simple_ class hierarchy, and attach the methods using said
template.