Fix type_methods default template params.
&type2t::type_id isn't a pointer into the derived type, it's a pointer
into type2t, so apparently isn't a valid param of type derived::*. Even
better, if we make it &derived::type_id, we end up with a cyclic
dependancy that g++ refuses to compile - the derived class derives from
type_methods, so the template potentially modifies its own parameters.

Fix this by inserting what in theory is a cast, but looks extremely
dodgy. Either way this compiles, and nothing depends on the field ptr
after the field type defaults, so no explosions should occur.