Fix schedule equation.
So a short time ago I added target_template, to allow for generating a
new target equation each time we start exploration. However, that was
being passed into the execution_statet object and having steps put into
it; rather than being kept as a template.

Instead, actually keep it as a template, and make the schedule target a
cloned copy of the template.