Fix a reachability situation;
So: in the past we've executed the execution guard after every sub-run
of a trace; after munging the symex stuff it was only executed when an
actual context switch occurs.

This turns out to be a bad plan - it means that from a certain point, if
we execute several sub-runs without context switching, all the
assumptions encoded in those sub-runs get guarded with a stale guard. To
illustrate:

    a    b     c    d    e    f
T1  o----o-----o----o----o----o
          \
           \
            \
             \ g    h    i    j
T2             o----o----o----o

We start some history at context switch point a, and execute the state
guard, lets call it G1. We then reach state b, and then _don't_ execute
the execution guard. From b there are then two histories: one where no
further context switches are taken (c, d, e, f), and where there's a
context switch to T2 (g, h, i j). If the context switch is taken, we
execute the state guard, to make G2.

The fault is that the assumptions and assigns in {a, b, c, d, e, f} are
all guarded with the first state guard, G1. That means that those
assumptions are still enforced in the second trace {a, b, g, h, i j},
making an overlap between these traces, which is bad.

So, execute the guard after every potential context switch instead.