Correct deadlock detection with joins.
A pthread join can detect that there's something else out there that's
blocked on a lock or condvar; however, it can't tell whether or not that
lock/condvar is unlocked/signaled when we come about joining.

So instead: maintain a count of threads blocked on joining. If we're
globally deadlocked, we can find an interleaving where one of the
blocked threads starts running, upon which it finds it isn't
locked/signalled and asserts. Otherwise, if it /is/ unlocked/signalled,
it just continues.

We assume that the target thread is ended after context switching away.
If it isn't, we avoid invalid behaviour through the assume. We can
probably find an interleaving where it's ended via some other means.
(cherry picked from commit cbc4f6797a8f9499298640d2045badabdb745c16)