Make another template change tack.
So: the previous arrangement of pulling in the relevant methods via
multiple inheretance doesn't fix the problem encountered, namely that we
end up with the address of the type2t in an irep2 not being the same as
the address for the object itself, breaking its use in shared_ptrs.

So try something slightly different: single inheretance, and the type
methods are a layer in between the top and anything that stores data.

The novelty here is the constructors - previously I didn't think there
was a way to pass down a variable number of type generic constructor
arguments that could be passed through the template to the underlying
object. However, it appears we can use the constructors in this patch as
they're only instanciated once something calls one of them. So, even
though type2t doesn't have a constructor that takes five
type2t::type_ids, because that type_methods constructor isn't called, it
isn't needed.

This may be a GCCism that we're suddenly depending on. In which case I
say colour me apathetic and uncaring; that just means we say you can
only compile with g++.