Explicit tmpl stuff; and template thrashing.
Add explicit template instanciations for bool_type2t, and empty_type2t.
Contrary to what I said in b207d013, this causes /all/ the constructors
in type_methods to be instanciated, search for the corresponding
constructor in subclass, and complain mightily that they don't exist.

So: abuse templates even further. By using the boost enable_if /
disable_if templates, we can enable or disable particular constructors
at a whim. Use this by ensuring that they're not enabled if one of their
arguments is going to be of type type2t::type_ids, because that'd
indicate that we'd be using a default (nonexistant) template parameter.

This requires that the constructor be a template itself. This is
unpleasent but necessary; what is supposed to happen is the constructor
becomes a template, and is disabled according to the type that
something is trying to instanciate. What actually happens is that we add
a dummy template ("arbitary") and disable or enable the whole thing
according to what the class template args are.

Enable or disable the single-type-id argument constructor according to
whether or not the subclass is type2t or not. Doesn't make sense
otherwise.

I forsee there'll be a problem when there's more than one field being
used, and the constructor for one field is enabled, but the subclass
we're using only has a two-field/argument constructor. This can be fixed
with more templating, which I shall do next.