Install more type-safety templating.
As referred to in 2db3ca8f, this attempts to prevent constructors from
being enabled when there are more enabled template arguments than it
handles. For example, if we have a type with 3 fields, the constructors
that take one and two of them shouldn't be enabled. Detect this by
seeing if the next field up is dummy, and if it isn't, disable.

Hasn't been tested; I'm hoping for the best.