Symbol_type2t -> new templates, do templ shuffle.
Move symbol type2t over to the latest template regeime. Here we discover
that you can't refer to fields in the derived/top-level type in
esbmct::type_methods due to that making a potentially recursive
definition. So, the data has to go in symbol_type_data, which gets
bunged elsewhere.

We also discover that G++ is extremely pedantic about the member-pointer
template parameter. To the extent that the expression _always_ has to be
&X::Y, and Y _has_ to be a member of X. It can't inheret it into X! This
breaks quite a few things in my default template parameters, and means
sharing fields across inhereted types becomes impossible.

Solve this by making everything even more type generic -  now
type_methods also take parameters for what class the field pointer it's
going to get is part of, which fixes the above problem. Happily it seems
we can apply base class member pointers to the derived type (as it
inherits), so we don't need to change the tempalte definitions that
much.