Move base_type_eq routines to irep2.
These have to maintain two kinds of routine, the old irep and the new
irep, because they're called from inside ansi-c as well as in symex
code. The ansi-c stuff can deal with a variety of incomplete types
(i.e., the "incomplete-struct" irep) and so breaks down if we try to
migrate that.

This isn't likely to change until we get to the entire code base being
non-old-irep, which is a long long way away.