Kill typecast operands to pointer_object2t.
These seem to have a habit to sneak their way into pointer_object exprs,
as a result of malloc returns being casted to unsigned ints. They also
result in type death on account of attempting to apply a tuple select to
something that's been casted to an integer.

So, get around this by ignoring typecasts when converting pointer_object
ireps. Same goes for pointer_offset.