Fix generation of nondeterministic arrays.
array_of(nondet) doesn't give the desired outcome - we end up with an
array of fields that are all initialized to the same value, wheras a
truly nondeterministic array has an array with all fields independantly
nondeterministic.

That and the problem that the Z3 backend, post irep2, spits out this
irep, as we can't cope with nondeterministically sized array_of's.
Seeing how SMT doesn't support it.