Don't assert initial access to argv.
It's disabled because the code used, to set the NULL terminator in argv,
is always correct. However, conversion code adds assertions to verify
that (because of nondeterminism). So, in the past, a "bounds_check"
field in an array access would stop that check being inserted.

However, that'd be placing a large amount of crap on top of irep2 if we
did the same thing, so instead actually check in the bounds_check
routine to see whether or not it's that initial access. This only
matches argv[argc], and only in the main function (due to the '
character in the name).

To prove that nothing else uses this feature, remove the relevant
accessors from the irep header. There could still be something elsewhere
using the string directly, in which case too bad, that's evidence of
old irep being terrible.