Start documenting ireps.
So the plan is this: document only the top level expressions/types, and
given them full documentation on the constructor methods, what the data
is used for, and so forth. Then don't document anything else, because it
won't all be copied onto one page by doxygen.

Readers will have to click through the class hierarchy to find out what
the field names are in the objects themselves; this is just too bad, and
I can't find a way around it.