Generate true and false constant exprs.
These are for locations where we're having to allocate a new constant
bool for true or false, which is unweildy.

This also involves setting some init priorities - the constructor for
constant_bool2t calls into type_pool, which has to be initialized first,
which means we run into the static initialization order fiasco. Fix this
by adding init priority numbers to ensure type_pool happens first.