Support llvm-data printing of assertions.
It would appear that the code wishes to extract a symbol name out of the
guard being used, without being bored about the actual guard operation
itself (and thus not caring about types :| ). Support this by migrating
back to old irep for this portion of code.

This can be squelched into being type safe sometime in the future.