Add additional library deps for Z3/libgomp
It demands pthreads and librt; this has me concerned, because linux
doesn't have a stable threading ABI, which means if we try to invoke
threads with a static binary on random kernel versions, it very much may
not work.

However, it would appear that Z3 doesn't attempt to use either OpenMP
and thus pthreads unless you explicitly tell it (and them maybe it
refuses on linux), so this isn't a problem right now.