Disable a thread-guard-is-false pruning feature.
So there's one thing I forgot: we may take a context switch with a false
thread guard (for some stupid reason). That context switch might be to
ourselves. When that happens, we disable ourselves because the guard is
false, even though by returning a couple of times we might end up with a
true guard again. So, disable this feature.