Start munging the symbol situation.
The plan for this branch is to move over symbol renaming to be a
non-string situation. i.e., have a symbol data structure that contains
integers representing how much something has been renamed. Rather than
STRINGS.

Start this off by making all code in ESBMC not access the member of
symbol2t directly, instead using an accessor method. Due to the sheer
amount of flailing back and forth between strings and irep_idts, this is
probably going to ruin performance for a bit.