Start seperating renaming data from strings.
Do this by extracting all the relevant data out of a string into the
symbol2t class when we construct one. Then when we pump it out into a
string, reconstruct it into the desired form.

Amazingly this appears to pass all tests that were passed before,
although I assume performance is being thrashed.