Fix renaming of global objects.
When something gets renamed up to level1, if it isn't in the level1
lookup tables then it gets nothing changed in its string. When it gets
renamed to level 2, it's in the global lookup table, so it always gets
renamed. Taht means a global object looks like

  somename&1#1

Without thread or stack frame identifiers. However this breaks with
irep2, as that isn't preserved; instead add additional enum types to
indicate that this behaviour should be honoured.