Add an irep2 symbol lookup api.
So, the renaming_nst has offended me. Unfortunately it's being passed
into the dereference code, which has otherly typed namespaces passed in
too. So, we can't just bulk-change the api.

Instead, add two ways of looking up symbols, with a symbol object or
with a string. We can then munge one kind of call into the other in the
future.

Ultimately the namespace object should only accept symbol2t lookups, and
assert that they're only renamed to level 1.