Remove irep_idt type from level2t storage.
Replace it with the level1 symbol object. So long as everything keeps
default values for the level2 values it's fine. Performance issues can
be fixed by using a hash map and caching hashes.

In theory there's no difference, but this should be a delightfully
non-stringy future.

The problem that might occur is tht now the type of a symbol is a hidden
part of the renaming data, that affects attempts to look up symbol.

Fallout: print functions in renaming::level2t become unconst.