Add class for recording a name entry...
... so basically, after mentioning that including a full symbol, and its
type, may lead to complications when looking up symbols, I've reached a
fundemental limitation.

Pointer type renaming is incosnsitent - some pointer types have a symbol
type hanging off them, others have an actual proper type in the subtype
field. That leads to some symbols with the same names that won't look up
correctly, and they can't be resolved without trying to make the pointer
type situation consistent, which is a vast undertaking.

So instead, define a class to contain something that's in the renaming
table. This is going to give us a vast amount more flexibility.