Add hash to name_record comparison methods.
This is important because the hash is a completely arbitary way to
measure two objects. One might be entirely greater than another but
still have a lesser hash.

So what this actually does is change the ordering of name_records in a
set to be in groups, ordered by the hash value, and within those groups
we define some ordering.