Fix a type situation.
This just popped up as 32_floppy failing - for some reason a call to
memcpy has been being stripped out, presumably by the slicer. In
226e7e9 I replaced the slicers internal storage mechamis, which changed
its behaviour, and as a result memcpy is now not cut out.

And as it turns out, memcpy is implemented by dereferencing as an array,
a void pointer. Which is invalid C; which isn't caught anywhere; which
leads to various dereference fails. Specifically invalid symbol objects
are created with the type "empty" (which is invalid) and if you give it
a default type, it turns out that array updates don't get the right
types in their operations.

So, fix the code, register ticket for fixing this.