Make local_variables record a name record set.
i.e., not a list of irep_idts.

Note that the fallout is changes to symex_valid_object.cpp, where I
can't generate a piece of equivalent code because there seems to be a
renaming level mismatch. That piece of code appears to be extremely
antiquated though, and it isn't killing any regression tests.