Add string container wrapper that hashes.
For storing hash maps of strings.

This is because I'm not convinced by reading the value set code that it
can't be easily turned into not indexing by strings without a
significant refactor. So, actually do store as strings, but don't use
irep_idt; instead store as a hash-mapped string wrapper.