Fix an execution guard renaming situation.
In 7d93863f, we made a symbol used by execution_state a symbol2t rather
than a string. However, it turns out to be using the wrong level of
renaming, as it triggers assertion failures during multithreaded tests.

Using a lower level of renaming works; I wonder whether or not dropping
a piece of information from the object is going to make anything break.
Unlikely, because all these things get renamed to be level2 unique
before being used anyway.