Add a wrapper for putting expr2tcs in hash maps
It turns out that what G++'s hash map does, is generate a hash lookup
table (which is all fine and well), and then iterates through a... list?
Comparing objects.  Which is slow for large objects.

So improve this by storing an expr2tc and a hash for it, and make the
comparison fast by first comparing the hash.