Revert "Add a wrapper for putting expr2tcs in hash maps"
This reverts commit a8f1b265788566deef5f9b2d4857bdf9823382c0.

So on closer inspection, it turns out that the encoding to Z3 is in fact
O(n^2 / 2). This is because for each expression tree, we hash-map-lookup
the whole thing, then when we don't find it we do the same thing to each
sub-node of the expression tree. So huge portions of the tree get hash'd
multiple times.

Faster lookups isn't going to fix this, a better way of encoding will.
That better way can be combined with not making the whole encoding
performed on the stack, because that's currently putitng a max size on
the expressions encoded.