Disable byte_extract on integer encoding mode.
Fundementally it doesn't make sense; there's no byte representation of
an unbounded integer. The only solution is to not to triger byte
representations of such integers, or to run in bitvector mode, two
things that are up to the user.

Some regression tests now abort; I'll fix this by changing encoding mode
default and any explicit options given.