Switch default Z3 encoding to bit-mode.
This is required to ensure that by default, any byte_extract and
byte_update ireps are actually representable, which isn't something that
can be done in integer mode.

This way, you can specify --z3-ir if you want, and it'll explode if you
do something funky with pointers. The default action is the most working
one.

Still broken are 03_nec07, 03_nec18, 03_nec37. These work in --z3-ir
mode; no idea why. Looks like the slicer and the byte extract/update
situation are in play. So, I'm going to leave them broken on the chance
further work will fix it.