Make pointer dereference offsets more accurate.
This results in the pointer offset (constant or dynamic) being added to
the offset resulting from the dereference. I.E., if we have:

  struct { int a; int b; int c; } face;
  struct face bees;
  struct face *knees = (struct face*)&bees.b;
  assert(knees->b == something);

This might not be valid in some strict aliasing situation that Denis
enjoys, but it represents something that might occur in reality, where
we have two offsets in play - the assignment to knees results in a
pointer with 4 bytes offset, and the dereference in the assert is an
additional 4 bytes.

In the modified code, the pointer offset from knees is already
calculated and passed into memory_model_bytes in the new_offset
argument. Then calculate_pointer_offset calculates the offset from the
dereference in the assertion.