Apply byte extract to the base data object.
That is, if we have a chain of indexes and member ireps applied to a
symbol, unwind them down to the base symbol.

That allows for actual accurate manipulation of data structures to
extract bytes from them. Unfortunately, it means assertion failures for
a while, because the Z3 backend still won't take it.