Fix encoding of solver questions
Firstly, we need to assert that all the in-program assumptions are true.

In addition, it turns out that l_not of a literal doesn't actually make
a new literal we can operate on, it just gives us the same literal via
an inversion. Not certain why I thought it would be any different.

That means if we set the literal to true or false, we then can't set it
to anything else without being guarenteed unsat. So instead push/pop
that assertion on; and seeing how we have to have two solver calls
anyway, ask directly for whether it can be true, whether it can be
false.