Clarify what kind of symbols mpor tracks.
The reason for l1 names is that l0 doesn't uniquely identify data in the
presence of recursion or reentrancy; l2 identifies each particular value
a variable can have and thus isn't useful. l1 uniquely identifies each
piece of data that might possibly receive a piece of memory to be stored
in, hence might result in a race.