Strip read/write sets from RT analysis path.
Currently we don't use it as we're not feeding an expr to POR right now.

Plus, the new idea is making all ART decisions outside of calls within
execution_statet. That neatly eliminates any problem with atomicity
within an instruction, because we only observe collected data after the
events and then decide to cswitch.

That then raises the question of how to handle ATOMIC_END's that should
force a context switch point; that requires extra data to be fed to ART
and more juggling to make sure POR doesn't strip it out?

I think the solution there is to essentially ignore them. During atomic
blocks, they'll be collecting symbols into their read/write set that
they access. But because cswitching is blocked, that'll never be
cleared. So the moment an atomic end happens, ART will be able to cause
a context switch because of the accessed fields. POR may still strip out
a context switch, but that'll be for some justified reason.

(Can't implement this yet because new POR isn't in place).