Correct thread dependancy checking.
Supposed to check dependancy through given thread l and the most
recently executed thread.

Next problem is that, by my reading of the MPOR paper, the "scheduling
constraint" data operates on the most recent transitions of each thread,
and "the transition of Ti enabled at time step k". I.E., it calculates
whether a thread is enabled based on whether it's _next_ transition is
dependant, etc.

Clearly that doesn't work for us because we don't know what the next
transition is going to be. So instead, I think the solution is to take
all transitions, and then look back and see whether or not we would have
been enabled in this circumstance.

However that takes a lot of juggling, so instead I'm going to change
tack and fix the ART stuff to only run after an instruction is complete,
not during.