Ileaves disabled doesn't stop thread being ended.
That is - in the previous arrangement, where a thread end was a context
switch point, an atomic block or the context switch count could prevent
execution stopping, which meant we executed past the end of the
thread. Which is bad.

Now, an ileave point and the end of the thread are two different things.

Things work now, we can return to POR.