Switch how mpor outcome is calculated.
As alluded to elsewhere, the arrangment where we can tell whether or not
a schedule is correct in advance isn't possible, because from my reading
we have to analyse the dependancy chain and /the transitions available
at this timestep/. Those transitions _aren't_ available and we don't
know what they might be, because we're symbolic executing stuff.

So instead of calcumalating MPOR data then deciding what to run, run
something then look back to see whether or not MPOR would have permitted
this.