Import MS z3++ header.
This file is copyright Microsoft and shipped out with Z3 4.0, however,
when we're compiling -Werror it causes all kinds of faults because gcc's
warning checking is stricter than whatever the Z3 team use to compile
their code with. So, it's going to have to live in-repository with a few
modifications to it.

I don't think it's a problem for licensing; if you use Z3 you're subject
to the academic free license because it's invoking their code. This is
no different than statically linking in the Z3 library itself.